Added Type.rb.  Used for type coercion (a-la Rubinius).
This is a ruby place holder so that we can start to have a common way to
try and do type coercion.  The long term solution is still under design,
but at least this solution allows us to gather together the places in the
code where the type coercion needs to happen (and some more mspecs should
pass).